moved firebase.js
