+++ app/src/store/index.js
@@ -1,6 +1,6 @@
 import Vue from "vue";
 import Vuex from "vuex";
-import * as fb from "../firebase";
+import * as fb from "../util/firebase";
 import router from "../router/index";
 
 Vue.use(Vuex);
